[#10] Extract LoginFormErrors into a sub-component. Adjust tests in a sub file
